Split session/parser modules under 700-line target
