+++ AGENTS.md
@@ -3,11 +3,17 @@
 ## Source Code Map
 - `src/lib.rs`: public database API surface and module exports.
 - `src/sql/mod.rs`: SQL subsystem wiring (lexer/parser/planner/executor/session).
-- `src/sql/parser/mod.rs`: SQL parser entry and core DDL/DML parsing helpers.
-- `src/sql/parser/expr_and_select.rs`: parser tail for set queries, expression parsing, and SELECT-body parsing.
+- `src/sql/parser/mod.rs`: SQL parser entry and statement dispatch.
+- `src/sql/parser/ddl_admin.rs`: parser DDL/admin statement parsing helpers.
+- `src/sql/parser/select_stmt.rs`: parser SELECT statement construction.
+- `src/sql/parser/query_common.rs`: parser shared query/update/delete parsing helpers.
+- `src/sql/parser/insert_stmt.rs`: parser INSERT/REPLACE statement parsing.
+- `src/sql/parser/expr_and_select.rs`: parser expression precedence and primary-expression parsing.
 - `src/sql/parser/tests.rs`: parser unit tests extracted from `parser/mod.rs`.
-- `src/sql/session/mod.rs`: runtime SQL session state machine (transaction lifecycle, checkpoint policy, stats SQL handlers).
+- `src/sql/session/mod.rs`: runtime SQL session state machine core (transaction lifecycle, execution routing).
+- `src/sql/session/checkpoint.rs`: session checkpoint policy/env parsing and stats/checkpoint handlers.
 - `src/sql/session/tests.rs`: session unit tests extracted from `session/mod.rs`.
+- `src/sql/session/tests/tail.rs`: session tests split from `tests.rs` (poison/read-only/rekey stats cases).
 - `src/sql/executor.rs` and `src/sql/executor/*.rs`: statement execution and query operators.
 - `src/sql/executor/tests.rs`: executor integration-style unit tests extracted from `executor.rs`.
 - `src/storage/pager/mod.rs`: pager core (header/cache/freelist IO, metadata accessors).
@@ -39,6 +45,9 @@
 - `src/fts/index.rs` was split into `src/fts/index/mod.rs` and `src/fts/index/tests.rs`.
 - `src/btree/ops.rs` was split into `src/btree/ops/mod.rs` and `src/btree/ops/tests.rs`.
 - `src/sql/parser.rs` was split into `src/sql/parser/mod.rs`, `src/sql/parser/expr_and_select.rs`, and `src/sql/parser/tests.rs`.
+- Parser responsibilities were further split into `src/sql/parser/ddl_admin.rs`, `src/sql/parser/select_stmt.rs`, `src/sql/parser/query_common.rs`, and `src/sql/parser/insert_stmt.rs`.
+- Session checkpoint/stats logic was split into `src/sql/session/checkpoint.rs`.
+- `src/sql/session/tests.rs` was further split with `src/sql/session/tests/tail.rs`.
 - `src/wal/recovery.rs` was split into `src/wal/recovery/mod.rs` and `src/wal/recovery/tests.rs`.
 - `src/sql/executor.rs` tests were split into `src/sql/executor/tests.rs`.
