+++ src/sql/session/tests/tail.rs
@@ -1,0 +1,258 @@
+use super::*;
+
+#[test]
+fn test_commit_in_doubt_increments_counter() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+    let wal_path = dir.path().join("test.wal");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    assert_eq!(session.database_stats().commit_in_doubt_count, 0);
+
+    // Set up a table first (needs a working pager)
+    session
+        .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
+        .unwrap();
+
+    // Inject write_page failure so next commit triggers CommitInDoubt
+    session
+        .pager_mut()
+        .set_inject_write_page_failure(Some(std::io::ErrorKind::Other));
+
+    let result = session.execute("INSERT INTO t VALUES (1)");
+    assert!(
+        matches!(&result, Err(MuroError::CommitInDoubt(_))),
+        "expected CommitInDoubt, got: {:?}",
+        result
+    );
+
+    // Counter must have been incremented
+    assert_eq!(session.database_stats().commit_in_doubt_count, 1);
+    assert!(session
+        .database_stats()
+        .last_commit_in_doubt_error
+        .is_some());
+    assert!(session
+        .database_stats()
+        .last_commit_in_doubt_timestamp_ms
+        .is_some());
+}
+
+#[test]
+fn test_show_checkpoint_stats_still_works() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    // SHOW CHECKPOINT STATS should still return 4 rows (backward compat)
+    match session.execute("SHOW CHECKPOINT STATS").unwrap() {
+        ExecResult::Rows(rows) => {
+            assert_eq!(rows.len(), 4);
+        }
+        _ => panic!("Expected rows from SHOW CHECKPOINT STATS"),
+    }
+}
+
+#[test]
+fn test_freelist_sanitize_stats_on_clean_open() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
+    let session = Session::new(pager, catalog, wal);
+
+    // Clean pager should have no sanitize stats
+    assert_eq!(session.database_stats().freelist_sanitize_count, 0);
+    assert_eq!(session.database_stats().freelist_out_of_range_total, 0);
+    assert_eq!(session.database_stats().freelist_duplicates_total, 0);
+}
+
+#[test]
+fn test_stats_readable_on_poisoned_session() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+    let wal_path = dir.path().join("test.wal");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    session
+        .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
+        .unwrap();
+
+    // Poison the session via CommitInDoubt
+    session
+        .pager_mut()
+        .set_inject_write_page_failure(Some(std::io::ErrorKind::Other));
+    let result = session.execute("INSERT INTO t VALUES (1)");
+    assert!(matches!(&result, Err(MuroError::CommitInDoubt(_))));
+
+    // Regular queries must be rejected
+    let result = session.execute("SELECT * FROM t");
+    assert!(matches!(&result, Err(MuroError::SessionPoisoned(_))));
+
+    // SHOW DATABASE STATS must still work on poisoned session
+    match session.execute("SHOW DATABASE STATS").unwrap() {
+        ExecResult::Rows(rows) => {
+            assert_eq!(rows.len(), 19);
+            // commit_in_doubt_count should be 1
+            assert_eq!(
+                rows[4].get("stat"),
+                Some(&Value::Varchar("commit_in_doubt_count".to_string()))
+            );
+            assert_eq!(rows[4].get("value"), Some(&Value::Varchar("1".to_string())));
+            // Pager cache stats should be present
+            assert_eq!(
+                rows[15].get("stat"),
+                Some(&Value::Varchar("pager_cache_hits".to_string()))
+            );
+            assert_eq!(
+                rows[16].get("stat"),
+                Some(&Value::Varchar("pager_cache_misses".to_string()))
+            );
+            assert_eq!(
+                rows[17].get("stat"),
+                Some(&Value::Varchar("pager_cache_hit_rate_pct".to_string()))
+            );
+        }
+        _ => panic!("Expected rows from SHOW DATABASE STATS"),
+    }
+
+    // SHOW CHECKPOINT STATS must also work on poisoned session
+    match session.execute("SHOW CHECKPOINT STATS").unwrap() {
+        ExecResult::Rows(rows) => {
+            assert_eq!(rows.len(), 4);
+        }
+        _ => panic!("Expected rows from SHOW CHECKPOINT STATS"),
+    }
+}
+
+#[test]
+fn test_rekey_wal_recreate_failure_poison_session() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+    let wal_path = dir.path().join("test.wal");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    session
+        .execute("CREATE TABLE t (id BIGINT PRIMARY KEY)")
+        .unwrap();
+    session.execute("INSERT INTO t VALUES (1)").unwrap();
+
+    session.inject_wal_recreate_failure_once_for_test();
+    let result = session.rekey_with_password("next_pass");
+    assert!(matches!(&result, Err(MuroError::SessionPoisoned(_))));
+
+    // Any regular statement is rejected after the WAL recreation failure.
+    let result = session.execute("SELECT * FROM t");
+    assert!(matches!(&result, Err(MuroError::SessionPoisoned(_))));
+
+    // Stats SQL remains available for operators.
+    match session.execute("SHOW DATABASE STATS").unwrap() {
+        ExecResult::Rows(rows) => {
+            assert_eq!(rows.len(), 19);
+        }
+        _ => panic!("Expected rows from SHOW DATABASE STATS"),
+    }
+}
+
+#[test]
+fn test_read_only_query_select_does_not_create_wal_records() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+    let wal_path = dir.path().join("test.wal");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&wal_path, &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    session
+        .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
+        .unwrap();
+    session
+        .execute("INSERT INTO t VALUES (1, 'alice')")
+        .unwrap();
+
+    let wal_before = std::fs::metadata(&wal_path).unwrap().len();
+    let rows = session
+        .execute_read_only_query("SELECT id, name FROM t")
+        .unwrap();
+    let wal_after = std::fs::metadata(&wal_path).unwrap().len();
+
+    assert_eq!(rows.len(), 1);
+    assert_eq!(rows[0].get("id"), Some(&Value::Integer(1)));
+    assert_eq!(wal_before, wal_after, "read-only query must not append WAL");
+}
+
+#[test]
+fn test_read_only_query_rejects_writes() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    let result = session.execute_read_only_query("INSERT INTO t VALUES (1)");
+    assert!(matches!(result, Err(MuroError::Execution(_))));
+}
+
+#[test]
+fn test_read_only_query_in_explicit_tx_reads_uncommitted_state() {
+    let dir = TempDir::new().unwrap();
+    let db_path = dir.path().join("test.db");
+
+    let mut pager = Pager::create(&db_path, &test_key()).unwrap();
+    let catalog = SystemCatalog::create(&mut pager).unwrap();
+    pager.set_catalog_root(catalog.root_page_id());
+    pager.flush_meta().unwrap();
+    let wal = WalWriter::create(&dir.path().join("test.wal"), &test_key()).unwrap();
+    let mut session = Session::new(pager, catalog, wal);
+
+    session
+        .execute("CREATE TABLE t (id BIGINT PRIMARY KEY, name VARCHAR)")
+        .unwrap();
+    session.execute("BEGIN").unwrap();
+    session
+        .execute("INSERT INTO t VALUES (1, 'alice')")
+        .unwrap();
+
+    let rows = session.execute_read_only_query("SELECT id FROM t").unwrap();
+    assert_eq!(rows.len(), 1);
+
+    session.execute("ROLLBACK").unwrap();
+    let rows = session.execute_read_only_query("SELECT id FROM t").unwrap();
+    assert!(rows.is_empty());
+}
